+++ scrapy/link.py
@@ -1,7 +1,7 @@
 """
 This module defines the Link object used in Link extractors.
 
-For actual link extractors implementation see scrapy.contrib.linkextractor, or
+For actual link extractors implementation see scrapy.linkextractors, or
 its documentation in: docs/topics/link-extractors.rst
 """
