+++ scrapy/linkextractors/regex.py
