+++ tests/test_engine.py
@@ -24,7 +24,7 @@ from scrapy.xlib.pydispatch import dispatcher
 from tests import tests_datadir
 from scrapy.spider import Spider
 from scrapy.item import Item, Field
-from scrapy.contrib.linkextractors import LinkExtractor
+from scrapy.linkextractors import LinkExtractor
 from scrapy.http import Request
 from scrapy.utils.signal import disconnect_all
